fix: move items into components folder, fix intergration
