+++ components/course/content/editor/MyAccordion.tsx
