+++ components/course/content/editor/MyTable.tsx
